+++ src/roiextractors/imagingextractor.py
@@ -1,25 +1,21 @@
 from abc import ABC, abstractmethod
-from typing import Union, Tuple
+from typing import Union, Optional, Tuple
 import numpy as np
 from copy import deepcopy
 
 from spikeextractors.baseextractor import BaseExtractor
 
-from .framesliceimaging import FrameSliceImaging
 from .extraction_tools import (
     ArrayType,
     PathType,
     DtypeType,
-    IntType,
     FloatType,
     check_get_videos_args,
 )
 
 
 class ImagingExtractor(ABC, BaseExtractor):
-    """An abstract class that contains all the meta-data and input data from
-    the imaging data.
-    """
+    """Abstract class that contains all the meta-data and input data from the imaging data."""
 
     def __init__(self):
         BaseExtractor.__init__(self)
@@ -133,7 +129,7 @@ class ImagingExtractor(ABC, BaseExtractor):
 
     def frame_slice(self, start_frame, end_frame):
         """Return a new ImagingExtractor ranging from the start_frame to the end_frame."""
-        return FrameSliceImaging(parent_imaging=self, start_frame=start_frame, end_frame=end_frame)
+        return FrameSliceImagingExtractor(parent_imaging=self, start_frame=start_frame, end_frame=end_frame)
 
     @staticmethod
     def write_imaging(imaging, save_path: PathType, overwrite: bool = False):
@@ -151,3 +147,57 @@ class ImagingExtractor(ABC, BaseExtractor):
             If True and save_path is existing, it is overwritten
         """
         raise NotImplementedError
+
+
+class FrameSliceImagingExtractor(ImagingExtractor):
+    """
+    Class to get a lazy frame slice.
+
+    Do not use this class directly but use `recording.frame_slice(...)`
+    """
+
+    extractor_name = "FrameSliceImagingExtractor"
+    installed = True
+    is_writable = True
+    installation_mesg = ""
+
+    def __init__(
+        self, parent_imaging: ImagingExtractor, start_frame: Optional[int] = None, end_frame: Optional[int] = None
+    ):
+        self._parent_imaging = parent_imaging
+        self._start_frame = start_frame
+        self._end_frame = end_frame
+        self._num_frames = self._end_frame - self._start_frame
+
+        parent_size = self._parent_imaging.get_num_frames()
+        if start_frame is None:
+            start_frame = 0
+        else:
+            assert 0 <= start_frame < parent_size
+        if end_frame is None:
+            end_frame = parent_size
+        else:
+            assert 0 < end_frame <= parent_size
+        assert end_frame > start_frame, "'start_frame' must be smaller than 'end_frame'!"
+
+        super().__init__()
+
+    def get_frames(self, frame_idxs: ArrayType, channel: int = 0) -> np.ndarray:
+        assert max(frame_idxs) < self._num_frames, "'frame_idxs' range beyond number of available frames!"
+        mapped_frame_idxs = np.array(frame_idxs) + self._start_frame
+        return self._parent_imaging.get_frames(frame_idxs=mapped_frame_idxs, channel=channel)
+
+    def get_image_size(self) -> Tuple[int, int]:
+        return tuple(self._parent_imaging.get_image_size())
+
+    def get_num_frames(self) -> int:
+        return self._num_frames
+
+    def get_sampling_frequency(self) -> float:
+        return self._parent_imaging.get_sampling_frequency()
+
+    def get_channel_names(self) -> list:
+        return self._parent_imaging.get_channel_names()
+
+    def get_num_channels(self) -> int:
+        return self._parent_imaging.get_num_channels()
